split code, added traits for better maintenance
